The shelfobject forms was moved from laboratory to shelfobject file
